+++ android/app/src/main/java/com/xianxia/sect/di/BridgeBindingsModule.kt
@@ -40,4 +40,7 @@ object BridgeBindingsModule {
 
     @Provides @Singleton
     fun provideHeavyDataDecoder(impl: HeavyDataDecoderImpl): HeavyDataDecoder = impl
+
+    @Provides @Singleton
+    fun provideWorldMapStatePort(impl: WorldMapStatePortImpl): WorldMapStatePort = impl
 }