+++ android/core/engine/src/main/java/com/xianxia/sect/core/GameEngine.kt
@@ -30,6 +30,7 @@ import com.xianxia.sect.core.engine.domain.save.SaveFacade
 import com.xianxia.sect.core.engine.service.HighFrequencyData
 import com.xianxia.sect.core.model.production.ProductionSlot
 import com.xianxia.sect.core.config.InventoryConfig
+import com.xianxia.sect.core.repository.WorldMapStatePort
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -81,7 +82,8 @@ class GameEngine @Inject constructor(
     internal val inventoryFacade: InventoryFacade,
     internal val diplomacyFacade: DiplomacyFacade,
     internal val productionFacade: ProductionFacade,
-    internal val saveFacade: SaveFacade
+    internal val saveFacade: SaveFacade,
+    internal val worldMapStatePort: WorldMapStatePort
 ) {
     companion object { private const val TAG = "GameEngine" }